feat(detection): 支持多算法置信度配置和截图优化
- 修改 ConfThreshold 类型从 float64为 float32 以提升性能
- 配置文件中 threshold 支持数组形式，可针对不同算法设置不同阈值
- transformer 中新增 ConfThreshold 字段并支持默认值
- 检测请求根据算法索引使用对应的置信度配置- 调整截图时间间隔配置从 200ms 到 5s
- 优化截图写入 OSS 流程，增强调试日志输出
- 图片绘制功能预留标签和置信度显示接口

+++ plugin/detection/pkg/algorithm.go
@@ -40,7 +40,7 @@ var AlgorithmMap = AlgorithmId{
 type DetectionRequest struct {
 	AlgorithmID   uint8   `json:"algorithm_id"`
 	Image         string  `json:"image"`
-	ConfThreshold float64 `json:"conf_threshold,omitempty"`
+	ConfThreshold float32 `json:"conf_threshold,omitempty"`
 }
 
 // DetectionResult 定义检测结果结构
